+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 - add satellite packaging machine (slightly nonfunctional)
 - make space not have the starting platform, rocket places it from the package.
+- fix GT compat.
+
+WARNING: space station doesn't build correctly just yet.